Adds magazines to loadouts (#3679)
# About the pull request

This pull request adds all of the various Boots! editions to the loadout
selection. I threw in a crayon and a hubba-hubba mag because I felt like
those were missing as well.

# Explain why it's good for the game


* Boots! is _THE_ USCM magazine. You'd expect there to be at least a
dozen fanboys on the Almayer carrying an edition or two.

* Crayons are sovlful. What else can I say? (I added the rainbow one
instead of a dozen different colors for ease)

* Hubba hubba.

# Testing Photographs and Procedure
Irrelevant.

# Changelog

:cl:
add: Adds various magazines and a crayon to loadout selection.
/:cl: